+++ pytket/pytket/circuit/decompose_classical.py
@@ -342,7 +342,7 @@ def _decompose_expressions(circ: Circuit) -> Tuple[Circuit, bool]:
             modified = True
             continue
         if optype == OpType.Barrier:
-            # add_gate doesn't work for metaops like barrier
+            # add_gate doesn't work for metaops
             newcirc.add_barrier(args)
         else:
             newcirc.add_gate(op, args, **kwargs)